% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_pip.R
\name{copy_dirs}
\alias{copy_dirs}
\title{Copy files from a folder}
\usage{
copy_dirs(dirs, output_path, input_path, ext = NULL)
}
\arguments{
\item{dirs}{Name of the directory to be copy}

\item{output_path}{where to create new folder}

\item{input_path}{file address to the PIP API data. Default is..}

\item{ext}{If only one extension is needed. Examples are "fst","qs" or "dta".}
}
\value{
TRUE
}
\description{
This function copy files using `fs::dir_copy` from and input folder to an
output folder and eliminates the `_vintage` folder (made explicitly to copy
`_aux` and `estimations` folders)
}
