% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_svy.R
\docType{data}
\name{datt}
\alias{datt}
\title{Group data from Datt (1998)}
\format{
A dataset with 7 variables:
\describe{
    \item{lwelfare}{Cumulative proportion of consumption}
    \item{welfare}{Mean monthly per capita consumption}
    \item{weight}{Cumulative proportion of population}
    \item{pop}{Percentage of persons}
    \item{country_code}{Country code for India}
    \item{reporting_level}{Reporting level: rural}
    \item{survey_year}{Year: 1983}
    }
}
\source{
Datt (1998) Computational tools for poverty measurement and analysis
}
\usage{
datt
}
\description{
Group data from Datt (1998)
}
\keyword{datasets}
