% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{del_files}
\alias{del_files}
\title{Delete files that do not have an specific extension}
\usage{
del_files(orig_file, ext_keep)
}
\arguments{
\item{orig_file}{File location}

\item{ext_keep}{extension to be kept in the folder}
}
\value{
none
}
\description{
Delete files that do not have an specific extension
}
