% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_svy.R
\docType{data}
\name{fk_cache_bin}
\alias{fk_cache_bin}
\title{Bin data (cache)}
\format{
A data frame with 33 variables:
\describe{
  \item{country_code}{Three letter country code}
  \item{survey_year}{Year of the survey}
  \item{surveyid_year}{Year of the survey (already rounded)}
  \item{survey_acronym}{Acronym of the survey (varies by country)}
  \item{hhid}{Household ID}
  \item{pid}{Person ID (not always available)}
  \item{welfare}{Welfare (consumption/income)}
  \item{weight}{Household weight (always?)}
  \item{gender}{Person gender (not always available)}
  \item{area}{Urban or rural area (not always available)}
  \item{welfare_type}{Type of welfare: consumption or income}
  \item{distribution_type}{micro}
  \item{cpi_data_level}{National, urban or rural data level used for CPI}
  \item{ppp_data_level}{National, urban or rural data level used for PPP}
  \item{pop_data_level}{National, urban or rural data level used for population}
  \item{gdp_data_level}{National, urban or rural data level used for GDP}
  \item{pce_data_level}{National, urban or rural data level used for PCE (?)}
  \item{gd_type}{??}
  \item{alt_welfare}{??}
  \item{imputation_id}{Imputation ID}
  \item{survey_id}{Survey ID}
  \item{vermast}{??}
  \item{veralt}{??}
  \item{collection}{??}
  \item{module}{??}
  \item{tool}{??}
  \item{source}{Source of the data: `BIN`}
  \item{reporting_level}{Reporting level of the survey}
  \item{ppp}{Purchase Power Parity}
  \item{cpi}{Consumper Price Index}
  \item{welfare_lcu}{Welfare in Local Currency Units (Same as welfare)}
  \item{welfare_ppp}{Welfare in PPP}
  \item{cache_id}{ID for specific survey in cache}
}
}
\source{
Fake data using a synthetic vector generated by random bin surveys
(see `fk_cache_bin_gen`).
}
\usage{
fk_cache_bin
}
\description{
A fake dataset with raw bin data as found in cache from
`//w1wbgencifs01/pip/pip_ingestion_pipeline/`
}
\keyword{datasets}
