% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_cache_imp_gen.R
\name{fk_cache_imputed_gen}
\alias{fk_cache_imputed_gen}
\title{Fake data generator for cache imputed}
\usage{
fk_cache_imputed_gen(pip_files, n_obs = NULL, n_sim = 50, seed_svy = 51089)
}
\arguments{
\item{pip_files}{character vector with all `cache_file` location from `pip_cache_inventory`.
Needs to be larger than 5 characters.}

\item{n_obs}{number of observation per imputation.}

\item{n_sim}{number of imputations. Default is 50.}

\item{seed_svy}{Seed for sampling of surveys from `cache_inventory`}
}
\value{
data.table
}
\description{
Fake data generator for cache imputed
}
