% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_cache_micro_gen.R
\name{fk_cache_micro_gen}
\alias{fk_cache_micro_gen}
\title{Fake data generator for cache micro}
\usage{
fk_cache_micro_gen(pip_files, svy_sample = 20, n_obs = NULL, seed_svy = 51089)
}
\arguments{
\item{pip_files}{character vector with all `cache_file` location from `pip_cache_inventory`.}

\item{svy_sample}{number of surveys to sample from `pip_cache_inventory`. Default is 20.}

\item{n_obs}{number of observation of data.table. Default is average of
observations of `svy_sample`}

\item{seed_svy}{Seed for sampling of surveys from `pip_cache_inventory`. Default is 51089.}
}
\value{
data.table
}
\description{
Fake data generator for cache micro
}
