% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_micro_gen.R
\name{fk_micro_gen}
\alias{fk_micro_gen}
\title{Fake data generator}
\usage{
fk_micro_gen(pip_inventory)
}
\arguments{
\item{pip_inventory}{generated by `pipload::pip_find_data`}
}
\value{
data.table
}
\description{
Fake data generator
}
