% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fk_uniq}
\alias{fk_uniq}
\title{Fake data.table with unique values from specific survey}
\usage{
fk_uniq(svy, n_obs)
}
\arguments{
\item{svy}{Selected survey}

\item{n_obs}{Number of observations, bins or quantiles}
}
\value{
list of variables
}
\description{
Fake data.table with unique values from specific survey
}
