% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_pip.R
\name{gen_welf}
\alias{gen_welf}
\title{Function to generate a fake welfare vector}
\usage{
gen_welf(fk_svy, w_vec, n_obs)
}
\arguments{
\item{fk_svy}{fake survey to attach the new vector}

\item{w_vec}{weighted welfare vector}

\item{n_obs}{number of observations of new vector}
}
\value{
data.base
}
\description{
Function to generate a fake welfare vector
}
