% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_remote_installed}
\alias{is_remote_installed}
\title{Test if specific branch is installed}
\usage{
is_remote_installed(package_name, remote_ref)
}
\arguments{
\item{package_name}{Name of the package you want to test}

\item{remote_ref}{Name of the branch you want to verify if it is installed}
}
\value{
alert
}
\description{
Test if specific branch is installed
}
