% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_files_cache}
\alias{load_files_cache}
\title{Load and merge data files from cache}
\usage{
load_files_cache(cache_file)
}
\arguments{
\item{cache_file}{a vector of file addresses
(usually `cache_file` variable from `cache_inventory`)}
}
\value{
list
}
\description{
Load and merge data files from cache
}
