% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_files_pip}
\alias{load_files_pip}
\title{Load data from files in PIP drive}
\usage{
load_files_pip(orig_file)
}
\arguments{
\item{orig_file}{a file address
(usually `orig` or `cache_file` variable from `cache_inventory`)}
}
\value{
list
}
\description{
Load data from files in PIP drive
}
