% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_svys}
\alias{load_svys}
\title{Load random surveys}
\usage{
load_svys(pip_files, seed_svy, svy_sample)
}
\arguments{
\item{pip_files}{character vector with all `cache_file` location from `pip_cache_inventory`.}

\item{seed_svy}{Seed for sampling of surveys from `pip_cache_inventory`. Default is 51089.}

\item{svy_sample}{number of surveys to sample from `pip_cache_inventory`. Default is 20.}
}
\value{
data.table
}
\description{
Load random surveys
}
