% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipfaker-package.R
\docType{package}
\name{pipfaker-package}
\alias{pipfaker}
\alias{pipfaker-package}
\title{pipfaker: Fake data for PIP packages}
\description{
This packages includes and generates fake data for different stages of the PIP pipeline. These are the types of data you can use...
}
\keyword{internal}
