% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_cache_group_gen.R
\name{fk_cache_group_gen}
\alias{fk_cache_group_gen}
\title{Fake data generator for cache group}
\usage{
fk_cache_group_gen(n_quantiles = 20)
}
\arguments{
\item{n_quantiles}{number of quantiles of data.table. Default is 20.}
}
\value{
data.table
}
\description{
The group data is calculated from a synthetic vector
generated using the Indian rural data from 1983 (Datt, 1998).
}
