% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_cache_imp_gen.R
\name{fk_cache_imputed_gen}
\alias{fk_cache_imputed_gen}
\title{Fake data generator for cache imputed}
\usage{
fk_cache_imputed_gen(n_obs = NULL, n_sim = 50, d_type = "D2", seed_svy = 51089)
}
\arguments{
\item{n_obs}{number of observation per imputation.}

\item{n_sim}{number of imputations. Default is 50}

\item{d_type}{Either `D1` or `D2`}

\item{seed_svy}{Seed for sampling of surveys from `cache_inventory`}
}
\value{
data.table
}
\description{
Fake data generator for cache imputed
}
