% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_micro_gen.R
\name{fk_micro_gen}
\alias{fk_micro_gen}
\title{Fake data generator}
\usage{
fk_micro_gen(pip_inventory, svy_sample = 20, n_obs = NULL, seed_svy = NULL)
}
\arguments{
\item{pip_inventory}{generated by `pipload::pip_find_data`}

\item{svy_sample}{number of surveys to sample from `pip_inventory`}

\item{n_obs}{number of observation of data.table. Default is average of
observations of `svy_sample`}

\item{seed_svy}{Seed for sampling of surveys from `pip_inventory`}
}
\value{
data.table
}
\description{
Fake data generator
}
