% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk_pip.R
\name{fk_pip}
\alias{fk_pip}
\title{Fake PIP data for API}
\usage{
fk_pip(output_path = NULL, input_path = NULL, n_obs = 400)
}
\arguments{
\item{output_path}{where to create new folder}

\item{input_path}{file address to the PIP API data. Default is..}

\item{n_obs}{observations for micro surveys. Default is 400.}
}
\value{
folder
}
\description{
This function recreates a fake (and smaller) version of the folder
needed for the PIP API either by using data from a specific path
(given by the user) or using the fixed fake data found in this package
}
