% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_files_pip}
\alias{load_files_pip}
\title{Load and merge data files from PIP drive}
\usage{
load_files_pip(orig_file)
}
\arguments{
\item{orig_file}{a vector of file addresses
(usually `orig` variable from `cache_inventory`)}
}
\value{
list
}
\description{
Load and merge data files from PIP drive
}
